+++ main.py
@@ -1,4 +1,4 @@
-from src.game_manager import GameManager
+from src.core.game_manager import GameManager
 
 if __name__ == "__main__":
     game = GameManager()