+++ src/__init__.py
@@ -1,15 +1,15 @@
 
 #! You can use that __init__.py file to make your imports cleaner in your main script
 
-from .player import Player 
-from .resource_item import ResourceItem
-from .inventory import Inventory
-from .game_manager import GameManager
-from .enemy import Enemy
-from .camera import Camera
-from .save_manager import SaveManager
-from .resource_node import ResourceNode
-from .action_manager import ActionManager
+from .entities.player import Player 
+from .entities.resource_item import ResourceItem
+from .systems.inventory import Inventory
+from .core.game_manager import GameManager
+from .entities.enemy import Enemy
+from .core.camera import Camera
+from .systems.save_manager import SaveManager
+from .entities.resource_node import ResourceNode
+from .systems.action_manager import ActionManager